The version number can now be specified

+++ install.py
@@ -1,28 +1,36 @@
 # Installs CodeIgniter to the current directory
 import urllib.request
+import urllib.request
 import os
 import sys
 import zipfile
 
-url = 'https://github.com/bcit-ci/CodeIgniter/archive/3.1.8.zip'
-zip_file = {
-    'name': 'temp.zip'
-}
+if len(sys.argv) > 1:
+    version_number = '3.1.8'
+    url = 'https://github.com/bcit-ci/CodeIgniter/archive/' + version_number + '.zip'
+    zip_file = {
+        'name': 'download.zip'
+    }
 
-print('Downloading file ', zip_file['name'], ' from ', url)
+    print('Downloading file', zip_file['name'], 'from', url)
 
-urllib.request.urlretrieve(url, zip_file['name'])
+    urllib.request.urlretrieve(url, zip_file['name'])
 
-print('Done')
-print('Extracting file contents...')
+    print('Done')
+    print('Extracting file contents...')
 
-zip_file['instance'] = zipfile.ZipFile(zip_file['name'], 'r')
-zip_file['instance'].extractall(sys.argv[1] or '')
-zip_file['instance'].close()
+    zip_file['instance'] = zipfile.ZipFile(zip_file['name'], 'r')
+    zip_file['instance'].extractall(sys.argv[1])
+    zip_file['instance'].close()
 
-print('Removing temprorary files...')
+    print('Removing temprorary files...')
 
-os.remove(zip_file['name'])
+    os.remove(zip_file['name'])
 
-print('Done.')
-print('Installation complete!')
+    print('Done.')
+    print('Installation complete!')
+else:
+    print('Bee Gold - CodeIgniter Installer')
+    print('MIT Callum John @ItsSeaJay 2018')
+    print('https://github.com/ItsSeaJay/bee-gold/')
+    print('Usage: python3 install.py [installation_path] [version_number]')
